refactor: move customization module into configuration, migrate to Pinia, rename wt-type-extension to field-extensions
- customization/ was a stray top-level Vuex module even though its router
  already mounted everything under configuration/customization/*; moved it
  to configuration/modules/customization to match that URL structure and
  the sibling lookups module's location.
- wt-type-extension renamed to field-extensions — the old name leaked the
  backend's wtTypes/typeExtensions API naming instead of describing what it
  actually is (custom fields on cases/contacts), unlike every sibling
  lookup module (custom-lookup, sources, priorities...).
- Both submodules migrated off Vuex onto createTableStore/createCardStore
  (@webitel/ui-datalist), following the lookup-pinia-store-refactor
  convention:
  - custom-lookups (admin CRUD for dictionary type definitions) now uses
    AdjunctTypesAPI directly; dropped the unused generatePermissionsApi
    wiring — the "administered" toggle was already hidden pending backend
    support, so there was nothing live to preserve, and object-level
    permissions have no Pinia equivalent yet (same gap blocking
    contact-groups).
  - field-extensions now uses WtTypeExtensionAPI directly; the shared
    `types-extensions/:id` route doesn't remount on repo change, so the
    card store is reset/initialized from an explicit route.params watcher
    instead of onMounted (same pattern as custom-lookup's :repo handling).
- Decoupled the shared field-definition editor (opened-custom-lookup-columns
  + custom-field-popup, moved default-value-input/type-field-select into a
  components/shared/ subfolder) from Vuex's useCardStore(namespace) — it now
  takes a plain itemInstance prop, so both the Pinia custom-lookups module
  and field-extensions can render it.
- Removed the old Vuex customization module registration from
  app/store/index.js; updated every cases/contacts consumer of
  useExtensionFields/display-dynamic-field-extension/wt-display-content to
  the new paths and names.

Depends on [webitel/webitel-ui-sdk#WTEL-9717] wiring up AdjunctTypesAPI/
WtTypeExtensionAPI and adding their validation schemas.

[WTEL-9717](https://webitel.atlassian.net/browse/WTEL-9717)
