+++ src/app/store/index.js
@@ -3,7 +3,6 @@ import { createStore } from 'vuex';
 
 import appearance from '../../modules/appearance/store/appearance';
 import configuration from '../../modules/configuration/store/configuration';
-import customization from '../../modules/customization/store/customization';
 
 export default createStore({
 	state: {
@@ -21,6 +20,5 @@ export default createStore({
 	modules: {
 		appearance,
 		configuration,
-		customization,
 	},
 });
