+++ src/modules/cases/components/cases-filters-panel.vue
@@ -19,7 +19,7 @@
 import { TableFiltersPanelComponent as TableFiltersPanel } from '@webitel/ui-datalist/filters';
 import { storeToRefs } from 'pinia';
 
-import { useExtensionFields } from '../../customization/modules/wt-type-extension/composable/useExtensionFields';
+import { useExtensionFields } from '../../configuration/modules/customization/modules/field-extensions/composables/useExtensionFields';
 import { filtersOptions } from '../configs/filtersOptions';
 import { CasesNamespace } from '../namespace';
 import { useCasesDatalistStore } from '../stores/datalist/casesDatalistStore';
