+++ src/modules/configuration/modules/customization/modules/custom-lookups/components/the-custom-lookups.vue
@@ -1,0 +1,253 @@
+<template>
+  <wt-page-wrapper
+    class="table-page"
+    :actions-panel="false"
+  >
+    <template #header>
+      <wt-page-header
+        :secondary-action="close"
+        hide-primary
+      >
+        <wt-breadcrumb :path="path" />
+      </wt-page-header>
+    </template>
+    <template #main>
+      <section class="table-section">
+        <header class="table-title">
+          <h3 class="table-title__title">
+            {{ t('objects.customLookup.customLookup', 2) }}
+          </h3>
+          <wt-action-bar
+            :include="[IconAction.ADD, IconAction.REFRESH, IconAction.DELETE]"
+            :disabled:add="!hasCreateAccess"
+            :disabled:delete="!selected.length || !hasDeleteAccess"
+            @click:refresh="loadDataList"
+            @click:add="add"
+            @click:delete="
+              askDeleteConfirmation({
+                deleted: selected,
+                callback: () => deleteMany(selected),
+              })
+              "
+          >
+            <template #search-bar>
+              <dynamic-filter-search
+                :filters-manager="filtersManager"
+                single-search-name="q"
+                @filter:add="addFilter"
+                @filter:update="updateFilter"
+                @filter:delete="deleteFilter"
+              />
+            </template>
+          </wt-action-bar>
+        </header>
+
+        <delete-confirmation-popup
+          :shown="isDeleteConfirmationPopup"
+          :callback="deleteCallback"
+          :delete-count="deleteCount"
+          @close="closeDelete"
+        />
+
+        <div class="table-section__table-wrapper">
+          <wt-empty
+            v-show="showEmpty"
+            :image="imageEmpty"
+            :headline="emptyHeadline"
+            :title="emptyTitle"
+            :text="textEmpty"
+            :primary-action-text="emptyPrimaryActionText"
+            @click:primary="add"
+          />
+
+          <wt-loader v-show="isLoading" />
+
+          <wt-table
+            v-show="dataList.length && !isLoading"
+            :data="dataList"
+            :headers="shownHeaders"
+            :selected="selected"
+            sortable
+            @sort="updateSort"
+            @update:selected="updateSelected"
+          >
+            <template #name="{ item }">
+              <wt-item-link :link="{
+                name: `${CrmSections.CustomLookups}-card`,
+                params: { id: item.repo },
+              }">
+                {{ item.name }}
+              </wt-item-link>
+            </template>
+            <template #about="{ item }">
+              {{ item.about }}
+            </template>
+            <template #createdAt="{ item }">
+              {{ prettifyDate(item.createdAt) }}
+            </template>
+            <template #createdBy="{ item }">
+              {{ item.createdBy?.name }}
+            </template>
+            <template #actions="{ item }">
+              <wt-icon-action
+                action="edit"
+                :disabled="!hasUpdateAccess"
+                @click="edit(item)"
+              />
+              <wt-icon-action
+                action="delete"
+                :disabled="!hasDeleteAccess"
+                @click="
+                  askDeleteConfirmation({
+                    deleted: [item],
+                    callback: () => deleteEls(item),
+                  })
+                  "
+              />
+            </template>
+          </wt-table>
+
+          <wt-pagination
+            :next="next"
+            :prev="page > 1"
+            :size="size"
+            debounce
+            @change="updateSize"
+            @next="updatePage(page + 1)"
+            @prev="updatePage(page - 1)"
+          />
+        </div>
+      </section>
+    </template>
+  </wt-page-wrapper>
+</template>
+
+<script lang="ts" setup>
+import { DynamicFilterSearchComponent as DynamicFilterSearch } from '@webitel/ui-datalist/filters';
+import { WtEmpty } from '@webitel/ui-sdk/components';
+import { useClose } from '@webitel/ui-sdk/composables';
+import { CrmSections, IconAction } from '@webitel/ui-sdk/enums';
+import DeleteConfirmationPopup from '@webitel/ui-sdk/src/modules/DeleteConfirmationPopup/components/delete-confirmation-popup.vue';
+import { useDeleteConfirmationPopup } from '@webitel/ui-sdk/src/modules/DeleteConfirmationPopup/composables/useDeleteConfirmationPopup';
+import { useTableEmpty } from '@webitel/ui-sdk/src/modules/TableComponentModule/composables/useTableEmpty';
+import { prettifyDate } from '@webitel/ui-sdk/utils';
+import { storeToRefs } from 'pinia';
+import { computed } from 'vue';
+import { useI18n } from 'vue-i18n';
+import { useRouter } from 'vue-router';
+
+import { useUserAccessControl } from '../../../../../../../app/composables/useUserAccessControl';
+import { useCustomLookupsDatalistStore } from '../stores/datalist/customLookupsDatalistStore';
+
+const { t } = useI18n();
+const router = useRouter();
+
+const { hasCreateAccess, hasUpdateAccess, hasDeleteAccess } =
+	useUserAccessControl();
+
+const tableStore = useCustomLookupsDatalistStore();
+
+const {
+	dataList,
+	selected,
+	error,
+	isLoading,
+	page,
+	size,
+	next,
+	shownHeaders,
+	filtersManager,
+} = storeToRefs(tableStore);
+
+const {
+	initialize,
+	loadDataList,
+	updateSelected,
+	updatePage,
+	updateSize,
+	updateSort,
+	deleteEls,
+	addFilter,
+	updateFilter,
+	deleteFilter,
+} = tableStore;
+
+initialize();
+
+const {
+	isVisible: isDeleteConfirmationPopup,
+	deleteCount,
+	deleteCallback,
+
+	askDeleteConfirmation,
+	closeDelete,
+} = useDeleteConfirmationPopup();
+
+const path = computed(() => [
+	{
+		name: t('crm'),
+		route: '/start-page',
+	},
+	{
+		name: t('startPage.configuration.name'),
+		route: '/configuration',
+	},
+	{
+		name: t('objects.customization.customization'),
+		route: '/configuration',
+	},
+	{
+		name: t('objects.customLookup.customLookup', 2),
+	},
+]);
+
+const { close } = useClose('configuration');
+
+function edit(item: { repo: string }) {
+	return router.push({
+		name: `${CrmSections.CustomLookups}-card`,
+		params: {
+			id: item.repo,
+		},
+	});
+}
+
+// This method for delete many lookups, one by one, because if we send many delete lookups requests at once, backend will return error
+const deleteMany = async (
+	items: Array<{
+		repo: string;
+	}>,
+) => {
+	for (const item of items) {
+		await deleteEls(item);
+	}
+};
+
+const {
+	showEmpty,
+	image: imageEmpty,
+	text: textEmpty,
+	headline: emptyHeadline,
+	title: emptyTitle,
+	primaryActionText: emptyPrimaryActionText,
+} = useTableEmpty({
+	dataList,
+	error,
+	filters: computed(() => filtersManager.value.getAllValues()),
+	isLoading,
+});
+
+const add = () => {
+	return router.push({
+		name: `${CrmSections.CustomLookups}-card`,
+		params: {
+			id: 'new',
+		},
+	});
+};
+</script>
+
+<style
+  lang="scss"
+  scoped
+></style>
