+++ src/modules/configuration/modules/customization/modules/custom-lookups/enums/FieldType.ts
