+++ src/modules/configuration/modules/customization/modules/custom-lookups/namespace.ts
@@ -1,0 +1,3 @@
+import { CustomizationNamespace } from '../../namespace';
+
+export const CustomLookupsNamespace = `${CustomizationNamespace}/customLookups`;
