+++ src/modules/configuration/modules/customization/modules/custom-lookups/stores/card/customLookupsCardStore.ts
@@ -1,0 +1,11 @@
+import { AdjunctTypesAPI } from '@webitel/api-services/api';
+import { createCardStore } from '@webitel/ui-datalist/card';
+
+import { CustomLookupsNamespace } from '../../namespace';
+import { customLookupsSchema } from '../../validations/customLookupsSchema';
+
+export const useCustomLookupsCardStore = createCardStore({
+	namespace: `${CustomLookupsNamespace}/card`,
+	apiModule: AdjunctTypesAPI,
+	standardValidationSchema: customLookupsSchema,
+});
