+++ src/modules/configuration/modules/customization/modules/field-extensions/namespace.ts
@@ -1,0 +1,3 @@
+import { CustomizationNamespace } from '../../namespace';
+
+export const FieldExtensionsNamespace = `${CustomizationNamespace}/fieldExtensions`;
