+++ src/modules/configuration/modules/customization/namespace.ts
@@ -1,0 +1,3 @@
+import { ConfigurationNamespace } from '../../namespace';
+
+export const CustomizationNamespace = `${ConfigurationNamespace}/customization`;
