+++ src/modules/contacts/modules/details/components/the-details.vue
@@ -28,7 +28,7 @@
         v-else
         class="opened-card-input-grid opened-card-input-grid--2-col opened-card-input-grid--w50"
       >
-        <wt-display-content
+        <field-extension-display
           v-for="field in fields"
           :key="field.id"
           :field="field"
@@ -48,9 +48,8 @@ import get from 'lodash/get';
 import { type ComputedRef, computed, inject, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { useRouter } from 'vue-router';
-
+import FieldExtensionDisplay from '../../../../configuration/modules/customization/modules/field-extensions/components/field-extension-display.vue';
 import CustomLookupDynamicField from '../../../../configuration/modules/lookups/modules/custom-lookup/components/custom-lookup-dynamic-field.vue';
-import WtDisplayContent from '../../../../customization/modules/wt-type-extension/components/wt-display-content.vue';
 import { useContactCardStore } from '../../../stores/card/contactCardStore';
 
 const access =
