--- src/modules/customization/modules/custom-lookups/components/default-value-input.vue
@@ -1,126 +1,0 @@
-<template>
-  <template v-if="displaySelect">
-    <wt-single-select
-      v-if="!multiple"
-      :key="lookup?.path + value.kind"
-      :model-value="value.default"
-      :options="options"
-      :label="t('customization.customLookups.defaultValue')"
-      required
-      :disabled="disabledDefaultValue"
-      :v="v.value.default"
-      :data-key="lookup?.primary || 'id'"
-      :search-method="loadLookupList"
-      @update:model-value="selectValue($event)"
-    />
-    <wt-multi-select
-      v-if="multiple"
-      :key="lookup?.path + value.kind"
-      :model-value="value.default"
-      :options="options"
-      :label="t('customization.customLookups.defaultValue')"
-      required
-      :disabled="disabledDefaultValue"
-      :v="v.value.default"
-      :data-key="lookup?.primary || 'id'"
-      :search-method="loadLookupList"
-      @update:model-value="selectValue($event)"
-    />
-  </template>
-  <template v-else-if="displayInput">
-    <wt-datepicker
-      v-if="kind === FieldType.Calendar"
-      :key="kind"
-      :label="t('customization.customLookups.defaultValue')"
-      v-model:model-value="value.default"
-      required
-      :v="v.value.default"
-      :disabled="disabledDefaultValue"
-      show-time
-    />
-    <wt-input-number
-      v-else-if="kind === FieldType.Number"
-      :key="kind"
-      v-model:model-value="value.default"
-      :disabled="disabledDefaultValue"
-      :label="t('customization.customLookups.defaultValue')"
-      required
-      :v="v.value.default"
-    />
-    <wt-input-text
-      v-else
-      :key="kind"
-      :disabled="disabledDefaultValue"
-      v-model:model-value="value.default"
-      :label="t('customization.customLookups.defaultValue')"
-      required
-      :v="v.value.default"
-    />
-  </template>
-</template>
-
-<script
-  setup
-  lang="ts"
->
-import { AdjunctTypeRecordsAPI } from '@webitel/api-services/api';
-import deepCopy from 'deep-copy';
-import { computed, ref, toRefs, watch } from 'vue';
-import { useI18n } from 'vue-i18n';
-import { FieldType } from '../enums/FieldType';
-import {
-	CustomLookupField,
-	CustomLookupLookup,
-	CustomLookupValue,
-} from '../types/customLookupField';
-
-const props = defineProps<{
-	value: CustomLookupField;
-	disabledDefaultValue: boolean;
-	// TODO(types): Vuelidate validation node; only the fields used in the
-	// template are typed here.
-	v: {
-		value: {
-			default: unknown;
-		};
-	};
-}>();
-
-const { lookup, kind } = toRefs(props.value);
-
-const { t } = useI18n();
-const options = ref([]);
-const displaySelect = computed(
-	() => kind.value === FieldType.Select || kind.value === FieldType.Multiselect,
-);
-const displayInput = computed(() => kind.value !== FieldType.Boolean);
-const multiple = computed(() => kind.value === FieldType.Multiselect);
-
-const getLoadLookupList = (lookup: CustomLookupLookup) => {
-	return (params) =>
-		AdjunctTypeRecordsAPI.getLookup({
-			...params,
-			path: lookup?.path,
-			display: lookup?.display || 'name',
-			primary: lookup?.primary || 'id',
-		});
-};
-const loadLookupList = ref(getLoadLookupList(lookup.value));
-
-const selectValue = (event: CustomLookupValue) => {
-	if (!Object.keys(event).length) {
-		props.value.default = null;
-		return;
-	}
-
-	props.value.default = deepCopy(event);
-};
-
-watch(
-	() => lookup.value,
-	(newValue) => {
-		props.value.default = null;
-		loadLookupList.value = getLoadLookupList(newValue);
-	},
-);
-</script>
