--- src/modules/customization/modules/custom-lookups/components/field-popup.vue
@@ -1,226 +1,0 @@
-<template>
-  <wt-popup
-    :shown="shown"
-    class="field-popup"
-    size="sm"
-    @close="close"
-  >
-    <template #header>
-      {{
-        !isNew
-          ? $t('customization.customLookups.editColumn')
-          : $t('customization.customLookups.addColumn')
-      }}
-    </template>
-    <template #main>
-      <div class="field-popup-wrapper" @input="changeDirty(true)">
-        <wt-input-text
-          v-model:model-value="value.name"
-          :label="$t('reusable.title')"
-          :v="v$.value.name"
-          required
-        />
-
-        <wt-input-text
-          v-model:model-value="value.id"
-          :label="t('customization.customLookups.code')"
-          :disabled="!isNew"
-          :v="v$.value.id"
-          :custom-validators="[
-            {
-              name: 'checkId',
-              text: $t('validation.latinWithNumber'),
-            },
-          ]"
-          required
-        />
-
-        <type-field-select
-          :disabled="!isNew"
-          :value="value"
-        />
-
-        <default-value-input
-          v-if="value.required"
-          :value="value"
-          :disabledDefaultValue="disabledDefaultValue"
-          :is-new="isNew"
-          :v="v$"
-        />
-
-        <wt-switcher
-          v-if="value.kind && value.kind !== FieldType.Boolean"
-          :model-value="value.required"
-          :label="$t('reusable.required')"
-          :v="v$"
-          @update:model-value="changeRequired($event)"
-        ></wt-switcher>
-      </div>
-    </template>
-    <template #actions>
-      <wt-button
-        :disabled="disabledSave"
-        @click="save"
-      >
-        {{ field ? $t('reusable.save') : $t('reusable.add') }}
-      </wt-button>
-      <wt-button
-        color="secondary"
-        @click="close"
-      >
-        {{ $t('reusable.close') }}
-      </wt-button>
-    </template>
-  </wt-popup>
-</template>
-
-<script setup>
-import { useVuelidate } from '@vuelidate/core';
-import { helpers, required } from '@vuelidate/validators';
-import deepCopy from 'deep-copy';
-import { computed, nextTick, ref, watch } from 'vue';
-import { useI18n } from 'vue-i18n';
-
-import { FieldType } from '../enums/FieldType.js';
-import DefaultValueInput from './default-value-input.vue';
-import TypeFieldSelect from './type-field-select.vue';
-
-const props = defineProps({
-	field: {
-		type: Object,
-		default: () => null,
-	},
-	shown: {
-		type: Boolean,
-		required: true,
-	},
-});
-
-const emit = defineEmits([
-	'save',
-	'close',
-]);
-
-const draft = {
-	name: '',
-	id: '',
-	kind: '',
-	required: false,
-	lookup: null,
-	list: null,
-	default: null,
-	_dirty: false,
-};
-const { t } = useI18n();
-
-const checkId = (repo) => {
-	const regex = /^[a-zA-Z_][a-zA-Z0-9_]*$/;
-	return regex.test(repo);
-};
-
-const value = ref(Object.assign(deepCopy(draft), deepCopy(props.field)));
-const isNew = computed(() => !props.field);
-
-const disabledDefaultValue = ref(!isNew.value);
-
-const updateValue = (newValue) => {
-	Object.assign(value.value, deepCopy(newValue));
-};
-
-const v$ = useVuelidate(
-	computed(() => ({
-		value: {
-			name: {
-				required,
-			},
-			id: {
-				required,
-				checkId: helpers.withMessage(t('validation.latinWithNumber'), checkId),
-			},
-			default: value.value.required
-				? {
-						required,
-					}
-				: {},
-		},
-	})),
-	{
-		value,
-	},
-	{
-		$autoDirty: true,
-	},
-);
-
-v$.value.$touch();
-
-const changeDirty = (dirty) => {
-	value.value._dirty = dirty;
-};
-
-const changeRequired = (event) => {
-	value.value.required = event;
-	value.value.default = null;
-	disabledDefaultValue.value = false;
-
-	nextTick(() => {
-		v$.value.$touch();
-	});
-};
-
-const disabledSave = computed(() => v$.value?.$invalid || !value.value._dirty);
-
-const save = () => {
-	const savedFiled = deepCopy(value.value);
-	delete savedFiled._dirty;
-
-	Object.keys(savedFiled).forEach((key) => {
-		if (!savedFiled[key] && key !== 'default') {
-			delete savedFiled[key];
-		}
-	});
-
-	emit('save', savedFiled);
-	changeDirty(false);
-	close();
-
-	if (!props.field) {
-		updateValue(draft);
-	}
-};
-
-const close = () => {
-	emit('close');
-};
-
-watch(
-	() => props.field,
-	() => {
-		if (props.field) {
-			updateValue(props.field);
-		}
-	},
-	{
-		deep: true,
-	},
-);
-
-watch(
-	() => props.shown,
-	(shown) => {
-		if (!shown && props.field) {
-			updateValue(props.field);
-		}
-	},
-);
-</script>
-
-<style lang="scss" scoped>
-.field-popup {
-  &-wrapper {
-    display: flex;
-    flex-direction: column;
-    gap: var(--spacing-sm);
-  }
-}
-</style>
