--- src/modules/customization/modules/custom-lookups/components/opened-custom-lookup.vue
@@ -1,176 +1,0 @@
-<template>
-  <wt-page-wrapper :actions-panel="!!currentTab.filters">
-    <template #header>
-      <wt-page-header
-        :primary-action="save"
-        :primary-disabled="disabledSave || !hasSaveActionAccess"
-        :primary-text="saveText"
-        :secondary-action="close"
-      >
-        <wt-breadcrumb :path="path" />
-      </wt-page-header>
-    </template>
-
-    <template #actions-panel>
-      <component
-        :is="currentTab.filters"
-        :namespace="namespace"
-      />
-    </template>
-
-    <template #main>
-      <form
-        class="main-container"
-        @submit.prevent="save"
-      >
-        <wt-tabs
-          :current="currentTab"
-          :tabs="tabs"
-          @change="changeTab"
-        />
-        <router-view v-slot="{ Component }">
-          <component
-            :is="Component"
-            :v="v$"
-            :is-new="isNew"
-            :namespace="cardNamespace"
-            :access="{
-              read: true,
-              update: !disableUserInput,
-              delete: !disableUserInput,
-              create: !disableUserInput,
-            }"
-          />
-        </router-view>
-        <input
-          hidden
-          type="submit"
-        />
-        <!--  submit form on Enter  -->
-      </form>
-    </template>
-  </wt-page-wrapper>
-</template>
-
-<script setup>
-import { useVuelidate } from '@vuelidate/core';
-import { helpers, required } from '@vuelidate/validators';
-import { CrmSections } from '@webitel/ui-sdk/enums';
-import { useCardComponent } from '@webitel/ui-sdk/src/composables/useCard/useCardComponent';
-import { useCardTabs } from '@webitel/ui-sdk/src/composables/useCard/useCardTabs';
-import { useClose } from '@webitel/ui-sdk/src/composables/useClose/useClose';
-import { useCardStore } from '@webitel/ui-sdk/src/store/new/index';
-import { computed } from 'vue';
-import { useI18n } from 'vue-i18n';
-
-import { useUserAccessControl } from '../../../../../app/composables/useUserAccessControl';
-import { useErrorRedirectHandler } from '../../../../error-pages/composable/useErrorRedirectHandler';
-
-const namespace = 'customization/customLookups';
-const { t } = useI18n();
-const { handleError } = useErrorRedirectHandler();
-
-const checkRepo = (repo) => {
-	const regex = /^[a-zA-Z_][a-zA-Z0-9_]*$/;
-	return regex.test(repo);
-};
-
-const {
-	namespace: cardNamespace,
-	id,
-	itemInstance,
-	...restStore
-} = useCardStore(namespace);
-
-const v$ = useVuelidate(
-	computed(() => ({
-		itemInstance: {
-			name: {
-				required,
-			},
-			repo: {
-				checkRepo: helpers.withMessage(
-					t('validation.latinWithNumber'),
-					checkRepo,
-				),
-			},
-		},
-	})),
-	{
-		itemInstance,
-	},
-	{
-		$autoDirty: true,
-	},
-);
-
-v$.value.$touch();
-
-const { isNew, pathName, saveText, save, initialize } = useCardComponent({
-	...restStore,
-	id,
-	itemInstance,
-	onLoadErrorHandler: handleError,
-});
-const { hasSaveActionAccess, disableUserInput } = useUserAccessControl();
-
-const { close } = useClose(CrmSections.CustomLookups);
-const disabledSave = computed(
-	() => v$.value?.$invalid || !itemInstance.value._dirty,
-);
-
-const tabs = computed(() => {
-	const general = {
-		text: t('reusable.general'),
-		value: 'general',
-		pathName: `${CrmSections.CustomLookups}-general`,
-	};
-	const columns = {
-		text: t('customization.customLookups.columns', 2),
-		value: 'columns',
-		pathName: `${CrmSections.CustomLookups}-columns`,
-	};
-
-	const tabs = [
-		general,
-	];
-
-	if (id.value) tabs.push(columns);
-	return tabs;
-});
-
-const { currentTab, changeTab } = useCardTabs(tabs);
-
-const path = computed(() => {
-	return [
-		{
-			name: t('crm'),
-			route: '/start-page',
-		},
-		{
-			name: t('startPage.configuration.name'),
-			route: '/configuration',
-		},
-		{
-			name: t('objects.customization.customization'),
-			route: '/configuration',
-		},
-		{
-			name: t('objects.customLookup.customLookup', 2),
-			route: {
-				name: CrmSections.CustomLookups,
-			},
-		},
-		{
-			name: isNew.value ? t('reusable.new') : pathName.value,
-		},
-	];
-});
-
-initialize();
-</script>
-
-<style
-  lang="scss"
-  scoped
-></style>
