--- src/modules/customization/modules/custom-lookups/components/the-custom-lookups.vue
@@ -1,259 +1,0 @@
-<template>
-  <wt-page-wrapper
-    class="table-page"
-    :actions-panel="false"
-  >
-    <template #header>
-      <wt-page-header
-        :secondary-action="close"
-        hide-primary
-      >
-        <wt-breadcrumb :path="path" />
-      </wt-page-header>
-    </template>
-    <template #main>
-      <section class="table-section">
-        <header class="table-title">
-          <h3 class="table-title__title">
-            {{ t('objects.customLookup.customLookup', 2) }}
-          </h3>
-          <wt-action-bar
-            :include="[IconAction.ADD, IconAction.REFRESH, IconAction.DELETE]"
-            :disabled:add="!hasCreateAccess"
-            :disabled:delete="!selected.length || !hasDeleteAccess"
-            @click:refresh="loadData"
-            @click:add="add"
-            @click:delete="
-              askDeleteConfirmation({
-                deleted: selected,
-                callback: () => deleteMany(selected),
-              })
-              "
-          >
-            <template #search-bar>
-              <filter-search
-                :namespace="filtersNamespace"
-                name="search"
-              />
-            </template>
-          </wt-action-bar>
-        </header>
-
-        <delete-confirmation-popup
-          :shown="isDeleteConfirmationPopup"
-          :callback="deleteCallback"
-          :delete-count="deleteCount"
-          @close="closeDelete"
-        />
-
-        <div class="table-section__table-wrapper">
-          <wt-empty
-            v-show="showEmpty"
-            :image="imageEmpty"
-            :headline="emptyHeadline"
-            :title="emptyTitle"
-            :text="textEmpty"
-            :primary-action-text="emptyPrimaryActionText"
-            @click:primary="add"
-          />
-
-          <wt-loader v-show="isLoading" />
-
-          <wt-table
-            v-show="dataList.length && !isLoading"
-            :data="dataList"
-            :headers="headers"
-            :selected="selected"
-            sortable
-            @sort="sort"
-            @update:selected="setSelected"
-          >
-            <template #name="{ item }">
-              <wt-item-link :link="{
-                name: `${CrmSections.CustomLookups}-card`,
-                params: { id: item.repo },
-              }">
-                {{ item.name }}
-              </wt-item-link>
-            </template>
-            <template #about="{ item }">
-              {{ item.about }}
-            </template>
-            <template #createdAt="{ item }">
-              {{ prettifyDate(item.createdAt) }}
-            </template>
-            <template #createdBy="{ item }">
-              {{ item.createdBy?.name }}
-            </template>
-            <template #actions="{ item }">
-              <wt-icon-action
-                action="edit"
-                :disabled="!hasUpdateAccess"
-                @click="edit(item)"
-              />
-              <wt-icon-action
-                action="delete"
-                :disabled="!hasDeleteAccess"
-                @click="
-                  askDeleteConfirmation({
-                    deleted: [item],
-                    callback: () => deleteData(item),
-                  })
-                  "
-              />
-            </template>
-          </wt-table>
-          <filter-pagination
-            :namespace="filtersNamespace"
-            :is-next="isNext"
-          />
-        </div>
-      </section>
-    </template>
-  </wt-page-wrapper>
-</template>
-
-<script setup>
-import { WtEmpty } from '@webitel/ui-sdk/components';
-import { CrmSections } from '@webitel/ui-sdk/enums';
-import { useClose } from '@webitel/ui-sdk/src/composables/useClose/useClose';
-import IconAction from '@webitel/ui-sdk/src/enums/IconAction/IconAction.enum';
-import DeleteConfirmationPopup from '@webitel/ui-sdk/src/modules/DeleteConfirmationPopup/components/delete-confirmation-popup.vue';
-import { useDeleteConfirmationPopup } from '@webitel/ui-sdk/src/modules/DeleteConfirmationPopup/composables/useDeleteConfirmationPopup';
-import FilterPagination from '@webitel/ui-sdk/src/modules/Filters/components/filter-pagination.vue';
-import FilterSearch from '@webitel/ui-sdk/src/modules/Filters/components/filter-search.vue';
-import { useTableFilters } from '@webitel/ui-sdk/src/modules/Filters/composables/useTableFilters';
-import { useTableEmpty } from '@webitel/ui-sdk/src/modules/TableComponentModule/composables/useTableEmpty';
-import { useTableStore } from '@webitel/ui-sdk/src/store/new/modules/tableStoreModule/useTableStore';
-import { prettifyDate } from '@webitel/ui-sdk/utils';
-import { computed, onUnmounted } from 'vue';
-import { useI18n } from 'vue-i18n';
-import { useRouter } from 'vue-router';
-
-import { useUserAccessControl } from '../../../../../app/composables/useUserAccessControl';
-
-const baseNamespace = 'customization/customLookups';
-
-const { t } = useI18n();
-const router = useRouter();
-
-const { hasCreateAccess, hasUpdateAccess, hasDeleteAccess } =
-	useUserAccessControl();
-
-const {
-	isVisible: isDeleteConfirmationPopup,
-	deleteCount,
-	deleteCallback,
-
-	askDeleteConfirmation,
-	closeDelete,
-} = useDeleteConfirmationPopup();
-
-const {
-	namespace,
-
-	dataList,
-	selected,
-	isLoading,
-	headers,
-	isNext,
-	error,
-
-	loadData,
-	deleteData,
-	sort,
-	setSelected,
-	onFilterEvent,
-} = useTableStore(baseNamespace);
-
-const {
-	namespace: filtersNamespace,
-	filtersValue,
-	restoreFilters,
-
-	subscribe,
-	flushSubscribers,
-} = useTableFilters(namespace);
-
-subscribe({
-	event: '*',
-	callback: onFilterEvent,
-});
-
-restoreFilters();
-
-onUnmounted(() => {
-	flushSubscribers();
-});
-
-/**
- * @author @Oleksandr Palonnyi
- *
- * [WTEL-7510](https://webitel.atlassian.net/browse/WTEL-7510)
- *
- * link for description https://webitel.atlassian.net/browse/WTEL-7510?focusedCommentId=693399
- * */
-const path = computed(() => [
-	{
-		name: t('crm'),
-		route: '/start-page',
-	},
-	{
-		name: t('startPage.configuration.name'),
-		route: '/configuration',
-	},
-	{
-		name: t('objects.customization.customization'),
-		route: '/configuration',
-	},
-	{
-		name: t('objects.customLookup.customLookup', 2),
-	},
-]);
-
-const { close } = useClose('configuration');
-
-function edit(item) {
-	router.push({
-		name: `${CrmSections.CustomLookups}-card`,
-		params: {
-			id: item.repo,
-		},
-	});
-}
-
-// This method for delete many lookups, one by one, because if we send many delete lookups requests at once, backend will return error
-const deleteMany = async (items) => {
-	for (const item of items) {
-		await deleteData(item);
-	}
-};
-
-const {
-	showEmpty,
-	image: imageEmpty,
-	text: textEmpty,
-	headline: emptyHeadline,
-	title: emptyTitle,
-	primaryActionText: emptyPrimaryActionText,
-} = useTableEmpty({
-	dataList,
-	error,
-	filters: filtersValue,
-	isLoading,
-});
-
-const add = () => {
-	router.push({
-		name: `${CrmSections.CustomLookups}-card`,
-		params: {
-			id: 'new',
-		},
-	});
-};
-</script>
-
-<style
-  lang="scss"
-  scoped
-></style>
