--- src/modules/customization/modules/wt-type-extension/components/the-wt-type-extension.vue
@@ -1,113 +1,0 @@
-<template>
-  <wt-page-wrapper
-    class="table-page"
-    :actions-panel="false"
-  >
-    <template #header>
-      <wt-page-header
-        :primary-action="save"
-        :primary-disabled="disabledSave"
-        :primary-text="saveText"
-        :secondary-action="close"
-      >
-        <wt-breadcrumb :path="path" />
-      </wt-page-header>
-    </template>
-
-    <template #main>
-      <form
-        class="main-container"
-        @submit.prevent="save"
-      >
-        <opened-custom-lookup-columns
-          :is-new="isNew"
-          :v="v$"
-          :title="$t(`customization.extensions.${repo}`)"
-          :namespace="cardNamespace"
-        />
-      </form>
-    </template>
-  </wt-page-wrapper>
-</template>
-
-<script setup>
-import { useVuelidate } from '@vuelidate/core';
-import { useCardComponent } from '@webitel/ui-sdk/src/composables/useCard/useCardComponent';
-import { useClose } from '@webitel/ui-sdk/src/composables/useClose/useClose';
-import { useCardStore } from '@webitel/ui-sdk/src/store/new/index';
-import { computed } from 'vue';
-import { useI18n } from 'vue-i18n';
-import { useRoute } from 'vue-router';
-
-import OpenedCustomLookupColumns from '../../custom-lookups/components/opened-custom-lookup-columns.vue';
-
-const namespace = 'customization/wtTypeExtension';
-
-const { t } = useI18n();
-const route = useRoute();
-
-const repo = computed(() => route.params.id);
-
-const {
-	namespace: cardNamespace,
-	id,
-	itemInstance,
-	...restStore
-} = useCardStore(namespace);
-
-const { isNew, saveText, save, initialize } = useCardComponent({
-	...restStore,
-	id,
-	itemInstance,
-});
-
-const { close } = useClose('customization');
-
-// This case need to get validation from child fields
-const v$ = useVuelidate(
-	{},
-	{
-		itemInstance,
-	},
-	{
-		$autoDirty: true,
-	},
-);
-
-v$.value.$touch();
-
-const disabledSave = computed(
-	() => v$.value?.$invalid || !itemInstance.value?._dirty,
-);
-
-/**
- * @author @Oleksandr Palonnyi
- *
- * [WTEL-7510](https://webitel.atlassian.net/browse/WTEL-7510)
- *
- * link for description https://webitel.atlassian.net/browse/WTEL-7510?focusedCommentId=693399
- * */
-const path = computed(() => {
-	return [
-		{
-			name: t('crm'),
-			route: '/start-page',
-		},
-		{
-			name: t('startPage.configuration.name'),
-			route: '/configuration',
-		},
-		{
-			name: t('objects.customization.customization'),
-			route: '/configuration',
-		},
-		{
-			name: repo.value && t(`customization.extensions.${repo.value}`),
-		},
-	];
-});
-
-initialize();
-</script>
-
-<style lang="scss" scoped></style>
